[refactor] Make `useCalculatedWidth` into `useCalculatedDimensions` and return height as well. (#12214)
## Describe your changes

When we are adding the new options for `height` to some elements like
vega/altair charts and graphviz charts, we need the calculated height as
well for the stretch height option. This refactors the existing
`useCalculatedWidth` hook that is used for these elements already to
also observe and return the calculated height.

<!-- If it's a visual change, please include a screenshot or video! -->

## GitHub Issue Link (if applicable)

## Testing Plan

- Unit Tests (JS and/or Python) ✅ 

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.