[refactor] Use useCalculatedDimensions in VegaLite.tsx (#12231)
## Describe your changes

This is a small change to bring this element up to date with the changes
for adding new width/height options to Vega lite charts.

<!-- If it's a visual change, please include a screenshot or video! -->

## GitHub Issue Link (if applicable)

## Testing Plan

Existing tests, no net changes in behaviour.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.