[refactor] Simplification of `st.graphviz_chart` rendering logic (#12296)
## Describe your changes

This logic was originally added [here to fix a bug
](https://github.com/streamlit/streamlit/pull/7398)with fullscreen
rendering. While working on height for `st.graphviz_chart`, I have found
it does not seem to be necessary any longer, I have not noticed the
described issues after removing the code. This may be because of the
changes that have been made to `StyledElementContainer` including the
min-width styling which was added to fix issues with returning from full
screen for other elements.

<!-- If it's a visual change, please include a screenshot or video! -->

## GitHub Issue Link (if applicable)

## Testing Plan

Existing tests. There is a snapshot test that covers the original bug. 

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.