[fix] `html` with float height.  (#12354)
## Describe your changes

The `html` element has width and height parameters which accept an
integer or None. Previously, we had no validation for the type and it
was possible to provide a float value. A recent change converted these
elements to the new width/height system and as part of that float values
were silently ignored and set to the default (stretch).

<!-- If it's a visual change, please include a screenshot or video! -->

## GitHub Issue Link (if applicable)

Closes https://github.com/streamlit/streamlit/issues/12340

## Testing Plan

I didn't add new tests for this because it isn't officially supported.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.

---------

Co-authored-by: lukasmasuch <lukas.masuch@gmail.com>