[fix] Share more code between Selectbox and Multiselect to align behavior (#12330)
## Describe your changes

- We previously fixed a bug in `st.selectbox`
(https://github.com/streamlit/streamlit/pull/12219), but not in
`st.multiselect`. Ultimately, this is because the code is effectively
duplicated between the 2, despite much of the behavior needing to be the
same.
- This PR moves the shared logic into a shared place, thus reducing
duplication and fixing the bug in both places as expected.

## GitHub Issue Link (if applicable)

## Testing Plan

- ✅ Updated unit tests

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.