Fix logo pushing out collapse button (#12329)
## Describe your changes

Fixes an issue with the logo pushing out the collapse button. This was
already working fine when a link was set -> the reason is that the
additional wrapping leads to a more correct calculation of the actual
available width. I think this is also solvable differently with some CSS
stuff, but I think its ok to just wrap the logo in a div when its not
already wrapped into a `a` element.

## GitHub Issue Link (if applicable)

- Closes https://github.com/streamlit/streamlit/issues/12326

## Testing Plan

- Updated the e2e tests to ensure that the collapse button is displayed
during snapshots. The logos in the previous version where somewhat
broken.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.