Use `key` as main identity for button widgets (#12332)
## Describe your changes

If a custom `key` is passed to any of the button widgets - `st.button`,
`st.download_button`, `st.form_submit_button` - it will be used as the
main identity. This allows for dynamically changing other parameters in
between reruns without causing the widgets to be recreated in the
frontend.

## GitHub Issue Link (if applicable)

- Closes https://github.com/streamlit/streamlit/issues/9402

## Testing Plan

- Added unit and e2e tests.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.