Add `sort` parameter to `st.bar_chart` (#12339)
## Describe your changes

This PR adds a `sort` parameter to `st.bar_chart` to sort the bars by a
column (`sort="col1"` for ascending and `sort="-col1"` for descending
order).

The default is `sort=True` for now, which keeps Altair's default
sorting. We might change this later on to disable sorting by default but
want to discuss it first.

## GitHub Issue Link (if applicable)

Closes #385
Closes #7111

## Testing Plan

- Added a few Python unit tests. 
- Added various snapshot tests for sorting. 

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.