Use `key` as main identity for `st.checkbox` & `st.toggle` (#12372)
## Describe your changes

If a custom `key` is passed to any of the `st.checkbox` / `st.toggle`
widgets it will be used as the main identity. This allows for
dynamically changing other parameters in between reruns without causing
the widgets to be recreated in the frontend.

## GitHub Issue Link (if applicable)

- Related to https://github.com/streamlit/streamlit/issues/12165

## Testing Plan

- Added e2e and unit tests.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.

---------

Co-authored-by: Lukas Masuch <103002573+sfc-gh-lmasuch@users.noreply.github.com>
Co-authored-by: graphite-app[bot] <96075541+graphite-app[bot]@users.noreply.github.com>