Use `key` as main identity for `st.time_input` & `st.date_input` (#12464)
## Describe your changes

If a custom `key` is passed to any of the `st.time_input` and
`st.date_input` widgets it will be used as the main identity. This
allows for dynamically changing other parameters in between reruns
without causing the widgets to be recreated in the frontend.

## GitHub Issue Link (if applicable)

- Related to https://github.com/streamlit/streamlit/issues/11277
- Closes https://github.com/streamlit/streamlit/issues/8410

## Testing Plan

- Added unit & e2e tests.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.