[chore] Modify Playwright changed files workflow (#12480)
## Describe your changes

<!-- If it's a visual change, please include a screenshot or video! -->

This PR stemmed from
https://github.com/streamlit/streamlit/pull/12473#pullrequestreview-3202063748

I noticed that the changed files workflow didn't fail the snapshots as I
expected. After talking with @lukasmasuch he said it's because it only
looks for changes to the _test.py files. This PR changes it so that it
will look at both.


## GitHub Issue Link (if applicable)

## Testing Plan

I tested it by changing the icon in st_alert.py and ensuring that the
workflow failed as expected.
<img width="1101" height="434" alt="image"
src="https://github.com/user-attachments/assets/6a0b16e2-9f58-4519-8623-facacac18dc1"
/>


---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.