Use `key` as main identity for custom components (#12472)
## Describe your changes

Refactors custom components to use the new `key_as_main_identity`
parameter instead of the previous approach. This doesn't change the
behaviour.

## GitHub Issue Link (if applicable)

- Related to https://github.com/streamlit/streamlit/issues/11277

## Testing Plan

- This has already been the behaviour, the change here is just a
refactoring to adapt it to how we do it in other widgets. The existing
tests should cover this fine.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.