[fix] Makefile help command delimiter to preserve :: in comments (#12564)
## Describe your changes

In pursuit of #12561 I ran into this bizarre issue where my :: kept
turning into a bunch of tabs. After some investigation I found the
culprit in this bit of bash here. I made the assumption that `:` is more
useful in make command helps than `;` is. While I was at it, claude
noticed that we have an extra space on the end of our command name so I
let it remove that with the `-1`

<!-- If it's a visual change, please include a screenshot or video! -->

## GitHub Issue Link (if applicable)

None

## Testing Plan

I ran the script.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.