[chore] Add make trace-e2e-test command for debugging e2e tests (#12561)
## Describe your changes

I wanted to use `npx playwright test --ui` but it seems this is only for ts/js tests. One downside of the python playwright decision is we don't get access to this. I remembered the stepping _back and forth_ UX I've used when developing Cypress tests. This is the closest thing I could create to that.

<!-- If it's a visual change, please include a screenshot or video! -->
<img width="1213" height="423" alt="image" src="https://github.com/user-attachments/assets/8183b15a-2cbf-459e-b3f1-70eaa911ca93" />
<img width="1277" height="798" alt="image" src="https://github.com/user-attachments/assets/0b214fc6-b27e-455b-92da-442fdc6e47b8" />



---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this project are made under the Apache 2.0 license.