Fix vega chart unrecognized dataset error (#11911)
## Describe your changes

Attempts to fix an issue with vega charts showing a blank chart with an
unrecognised data set error in the dev console. This is caused by two
race conditions in the current implementation.

## GitHub Issue Link (if applicable)

- Closes https://github.com/streamlit/streamlit/issues/11342
- Closes https://github.com/streamlit/streamlit/issues/11906

## Testing Plan

- Added unit tests.
- Since replicating this is not deterministic, its a bit hard to write
proper e2e tests.
---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.